+++ exporter/opentelemetry-exporter-zipkin/CHANGELOG.md
@@ -9,6 +9,8 @@
   ([#1064](https://github.com/open-telemetry/opentelemetry-python/pull/1064))
 - Zipkin exporter report instrumentation info. 
   ([#1097](https://github.com/open-telemetry/opentelemetry-python/pull/1097))  
+- Drop support for Python 3.4
+  ([#1099](https://github.com/open-telemetry/opentelemetry-python/pull/1099))
 
 ## Version 0.12b0